+++ s/signals/index.ts
@@ -2,7 +2,7 @@
 export * from "./parts/reactive.js"
 export * from "./parts/readable.js"
 
-export * from "./derive.js"
+export * from "./derived.js"
 export * from "./effect.js"
 export * from "./porcelain.js"
 export * from "./lazy.js"